pyproject.toml
[tool.poetry]
name = "mastermind"
version = "0.1.0"
description = "API project for playing mastermind"
authors = ["Javier Aguilera <jaguilerapuerta@gmail.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-extensions = "^3.1.5"
Werkzeug = "^2.0.3"
exrex = "^0.10.5"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-django = "^4.5.2"
pylint = "^2.12.2"
ipython = "^8.0.1"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
